+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [1.94.5](https://github.com/chris-yyau/busdriver/compare/v1.94.4...v1.94.5) (2026-07-19)
+
+
+### Bug Fixes
+
+* **pr-grind:** fail-CLOSED worktree resolution in Step 0 ([#421](https://github.com/chris-yyau/busdriver/issues/421)) ([#430](https://github.com/chris-yyau/busdriver/issues/430)) ([6a861fd](https://github.com/chris-yyau/busdriver/commit/6a861fd5fa40bbf04f341b1b70cb81bc5b902a67))
+
+## [1.94.4](https://github.com/chris-yyau/busdriver/compare/v1.94.3...v1.94.4) (2026-07-19)
+
+
+### Bug Fixes
+
+* **pr-grind:** poll for Codex first engagement instead of a blind 20s sleep ([#420](https://github.com/chris-yyau/busdriver/issues/420)) ([#428](https://github.com/chris-yyau/busdriver/issues/428)) ([81003cb](https://github.com/chris-yyau/busdriver/commit/81003cb883db307e173e0fb53c72987598624413)), closes [409/#390](https://github.com/chris-yyau/busdriver/issues/390) [#419](https://github.com/chris-yyau/busdriver/issues/419) [#413](https://github.com/chris-yyau/busdriver/issues/413) [409/#390](https://github.com/chris-yyau/busdriver/issues/390) [#413](https://github.com/chris-yyau/busdriver/issues/413)
+
 ## [1.94.3](https://github.com/chris-yyau/busdriver/compare/v1.94.2...v1.94.3) (2026-07-18)